Removed some extra loaders in webpack and reverted TS back to 4.2.2
