+++ package.json
@@ -88,7 +88,7 @@
     "stylelint-config-standard": "^22.0.0",
     "ts-loader": "^9.2.3",
     "tslib": "2.1.0",
-    "typescript": "4.3.2",
+    "typescript": "4.2.2",
     "webpack": "^5.38.1",
     "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^3.11.2",
